parallel brute force solver mostly working
parallel brute force solver is now fully functional for runs where the number of valid cell values at the first recursion level is <= the total number of ranks. A small amount of additional logic is necessary to allow the brute force solver to function correctly regardless of the total number of ranks, by continuing the initial traversal check through as many recursion levels as necessary.